feat: enhance signoff check by adding message body and a better regex (#141)
* commit: Include the body when signoff is being checked

The current implementation of signoff policy check is considering only
the subject. This is wrong because the signoff line is added in the body
of a commit message.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* Replace `get_commits_info` by `get_commit_info`

This name of the function is actualy wrong because it fetches only one
single commit. I also added a sha parameter to help in the future to
fetch older commits from the tree. Then, we can analyse multiple commits
as the original function name suggests.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* regex: Rebase the signoff regex to match the email too

Before this commit, the signoff regex only looks for the exact word.
This new proposal also checks if the commit has an email associated to
it.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* tests: Update mock call of `get_commit_info` command

This commit just updates the command line that is called by
`get_commit_info` to reflect the latest changes.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* fix: update re.match to re.search to fix test failure

---------

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Co-authored-by: shenxianpeng <xianpeng.shen@gmail.com>